project/requirements.txt
spotipy
scikit-learn==1.4.0
kaggle
python-dotenv
pandas